+++ library/src/main/java/com/project/library/models/Autor.java
@@ -1,10 +1,8 @@
-
-
 package com.project.library.models;
 
 public class Autor {
 
-    private int id;
+    private Long id;
     private String nome;
     private String nacionalidade;
 
@@ -15,11 +13,11 @@ public class Autor {
         this.nacionalidade = nacionalidade;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
